# openapi.yaml
openapi: 3.0.1
info:
  title: AkiraStack Auth
  description: 'This is the API Documentation'
  version: 1.0.0
...
servers:
- url: 'https://continuous-security.neuralcompany.team/api/auth'
paths:
...
  /v1/verifymail:
    post:
      tags:
      - signup
      description: "account verification through email"
      summary: "verify account using email link"
      operationId: verifyMail
      requestBody:
        description: "enter email and verification token"
        content:
          application/json:
            schema:
              properties:
                verification_code:
                  type: string
                email:
                  type: string
      responses:
        '200':
          description: "email verified"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/verification'
        '400':
          description: "bad request"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: boolean
                  message:
                    type: string
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: boolean
                  message:
                    type: string
...
  /v1/forgotpass:
    post:
      tags:
      - login
      summary: "send forgot password request and reset password using token"
      operationId: forgotPass
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              properties:
                state:
                  type: string
                  example: "forgot/reset"
                email:
                  type: string
                  example: "pass when type is forgot"
                token:
                  type: string
                  example: "pass when type is reset"
                password:
                  type: string
                  example: "pass when type is reset"
      responses:
        '200':
          description: "forgot-password/reset-password successful"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: boolean
                  message:
                    type: string

        '400':
          description: "invalid or missing data"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: boolean
                  message:
                    type: string
...
  /v1/google/login:
    get:
      tags:
      - Social Login
      summary: "login using google account"
      operationId: GoogleLogin
      parameters:
      - name: redirect
        in: query
        description: "redirect url after successfull login"
        schema:
          type: string
      responses:
        '302':
          description: "redirect to the url passed in query data with query data token(JWT token), name and email of user"
components:
  schemas:
...
    verification:
      type: object
      properties:
        email:
          type: string
        error:
          type: boolean
        expire:
          type: string
        message:
          type: string
        name:
          type: string
        role_id:
          type: string
        token:
          type: string
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT